Only reading in Poem
I noticed that the entirety of the poem was the only thing inside <p> tags.
I skip all words not in paragraph tags and then parse the ones that are in between paragaph tags.
Additionally, all words are converted to lowercase and all periods, commas, and explamation points are
removed from the end of the word to combine as many like-words as possible.